Organise endpoint structure.
Add sample packages to complete idea.
