+++ muster-kt/src/main/kotlin/com/muster/kt/endpoints/ministry/saveministry/CreateMinistryRequest.kt
@@ -1,3 +1,3 @@
-package com.muster.kt.endpoints
+package com.muster.kt.endpoints.ministry.saveministry
 
 data class CreateMinistryRequest(val name: String, val description: String)
