+++ muster-kt/src/main/kotlin/com/muster/kt/endpoints/ministry/saveministry/MinistryController.kt
@@ -1,6 +1,5 @@
-package com.muster.kt.endpoints
+package com.muster.kt.endpoints.ministry.saveministry
 
-import com.muster.kt.model.Ministry
 import org.http4k.core.Response
 import org.http4k.core.Status
 import java.util.*
