+++ muster-kt/src/test/kotlin/com/muster/kt/endpoints/ministry/saveministry/MinistryControllerImplTest.kt
@@ -1,6 +1,5 @@
-package com.muster.kt.endpoints
+package com.muster.kt.endpoints.ministry.saveministry
 
-import com.muster.kt.model.Ministry
 import com.natpryce.hamkrest.assertion.assertThat
 import org.http4k.core.Status
 import org.http4k.hamkrest.hasBody
